+++ CHANGELOG.md
@@ -22,6 +22,7 @@ Changelog for NeoFS Node
 - Use stream API of FSTree for object service `Get` operation (#3466, #3568)
 - Use meta buckets to mark containers with GC (#3561)
 - Switched to local BoltDB fork based on go.etcd.io/bbolt version 1.4.3 (#3576)
+- Move processing of expired objects from the epoch event handler to the regular GC cycle (#3582)
 
 ### Removed
 - `neofs-cli object head --main-only` no-op flag (#3509)